move the vc invite declined notification logic to space admin
