+++ src/common/enums/notification.event.ts
@@ -20,6 +20,7 @@ export enum NotificationEvent {
   SPACE_ADMIN_COLLABORATION_CALLOUT_CONTRIBUTION = 'SPACE_ADMIN_COLLABORATION_CALLOUT_CONTRIBUTION',
   SPACE_LEAD_COMMUNICATION_MESSAGE = 'SPACE_LEAD_COMMUNICATION_MESSAGE',
   SPACE_ADMIN_COMMUNITY_NEW_MEMBER = 'SPACE_ADMIN_COMMUNITY_NEW_MEMBER',
+  SPACE_ADMIN_VIRTUAL_CONTRIBUTOR_COMMUNITY_INVITATION_DECLINED = 'SPACE_ADMIN_VIRTUAL_CONTRIBUTOR_COMMUNITY_INVITATION_DECLINED',
   SPACE_COMMUNITY_INVITATION_USER_PLATFORM = 'SPACE_COMMUNITY_INVITATION_USER_PLATFORM',
   SPACE_COMMUNICATION_UPDATE = 'SPACE_COMMUNICATION_UPDATE',
   SPACE_COLLABORATION_CALLOUT_PUBLISHED = 'SPACE_COLLABORATION_CALLOUT_PUBLISHED',
@@ -39,7 +40,6 @@ export enum NotificationEvent {
 
   // Virtual contributors
   VIRTUAL_CONTRIBUTOR_ADMIN_SPACE_COMMUNITY_INVITATION = 'VIRTUAL_CONTRIBUTOR_ADMIN_SPACE_COMMUNITY_INVITATION',
-  VIRTUAL_CONTRIBUTOR_ADMIN_SPACE_COMMUNITY_INVITATION_DECLINED = 'VIRTUAL_CONTRIBUTOR_ADMIN_SPACE_COMMUNITY_INVITATION_DECLINED',
 }
 
 registerEnumType(NotificationEvent, {
